+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [X.X.X]
+* Feature:
+	* Add a public method to update now playing info metadata
+
 ---
 ## [0.9.3]
 * Feature: